% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{AB}
\alias{AB}
\title{AB}
\usage{
AB(ab)
}
\arguments{
\item{ab}{A vector of longform antibiotic names}
}
\value{
A vector of shortform names for antibiotics
}
\description{
Translate antibiotic names to shortform
}
