% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{AntibioticClass}
\alias{AntibioticClass}
\title{AntibioticClass}
\usage{
AntibioticClass(ab)
}
\arguments{
\item{ab}{A vector of longform antibiotic names}
}
\value{
The class of antibiotic
}
\description{
Classify AB into functional groups
}
