% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{collapse_AMR}
\alias{collapse_AMR}
\title{collapse_AMR}
\usage{
collapse_AMR(df_AMR)
}
\arguments{
\item{df_AMR}{the dataobject wity AMR at the isolate level}
}
\value{
a data.table object
}
\description{
Summarize AMR data at the sample level Sort in order to keep the
fingerprints consistent
}
