% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{EUvet}
\alias{EUvet}
\title{EUvet}
\usage{
EUvet(ab)
}
\arguments{
\item{ab}{A vector of longform antibiotic names}
}
\value{
The class in A B C D where A is the most important
}
\description{
Classify AB into importance according to EUvet
}
