% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.R
\name{figure1}
\alias{figure1}
\title{figure1}
\usage{
figure1(df_prev = prev_by_samplingID(), years = c(2023, 2024), check = FALSE)
}
\arguments{
\item{df_prev}{The data object}

\item{years}{the years to filter}

\item{check}{FALSE by default. If TRUE then the function returns
the processed data for testing purposes.}
}
\value{
A plot
}
\description{
Produce Figure 1 in the main chapter
}
