% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{host_association}
\alias{host_association}
\title{host_association}
\usage{
host_association(spa)
}
\arguments{
\item{spa}{The spa type}
}
\value{
A vector of the host association
}
\description{
estimate the host association of the SPA type
}
