% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{prev_by_samplingID}
\alias{prev_by_samplingID}
\title{prev_by_samplingID}
\usage{
prev_by_samplingID(df_prev = read_prev())
}
\arguments{
\item{df_prev}{the prevlance dataset}
}
\value{
An aggrigated dataset by samplingID
}
\description{
Aggregate by sampling ID. This is the YEAR, MATRIX_C, SAMPUNIT,
REPCOUNTRY, SAMPAREA, TOTUNITSTESTED and TOTUNITSPOSITIVE.
}
