% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{ST2CC}
\alias{ST2CC}
\title{ST2CC}
\usage{
ST2CC(ST)
}
\value{
A vector of infered CC from ST
}
\description{
STs 9 and 8325 were classified as CC1 by PubMLST1.
ST22 was classified as CC22 by PubMLST1.
}
