% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{table1}
\usage{
table1(spaprev = prev_by_SPA(), path_csv = tempfile(fileext = ".csv"))
}
\arguments{
\item{spaprev}{The object from the prev_by_SPA() function}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Table 1 - CC398 only in main chapter
Produce Table 1 in the main chapter
}
