% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_2.R
\name{table123_inner}
\alias{table123_inner}
\title{table123_inner}
\usage{
table123_inner(data = read_prev(), path_csv)
}
\arguments{
\item{data}{the data from read_prev()}

\item{path_csv}{path to the csv file}
}
\value{
A table
}
\description{
table123_inner
}
