% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE1_E2.R
\name{table1_2_inner}
\alias{table1_2_inner}
\title{table1_2_inner}
\usage{
table1_2_inner(
  df_prev = prev_by_samplingID(),
  year = NULL,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to a csv file}
}
\value{
a file path
}
\description{
table1_2_inner
}
