% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.R
\name{table3}
\alias{table3}
\title{table3}
\usage{
table3(df_AMR = read_AMR(), path_csv = tempfile(fileext = ".csv"))
}
\arguments{
\item{df_AMR}{The isolate based data object}

\item{path_csv}{path to the output csv file}

\item{years}{the years to filter}
}
\value{
A path to a csv file
}
\description{
Table 3 - Those with undefined CC (ie NA) - OBS this table is not
included in the chapter but is useful.
}
