% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE1_E2.R
\name{tableE1}
\alias{tableE1}
\title{tableE1}
\usage{
tableE1(
  df_prev = prev_by_samplingID(),
  year = 2024,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce Annex E Table 1: Methicillin-resistant Staphylococcus
aureus in food, 2024.
}
\details{
In the 2024 data as of 2025-12-10 there is a mistake in the German
reports in the prevlance data. This has resulted in an incorrect
numerator (number of positives) for the German: broiler meat,
turkey meat and crustaceans. In order to correct this we have used
these three numbers from the isolate based data like this:

foo <- collapse_AMR(read_AMR())
setDT(foo)
foo[year == 2024,
    .(pos = sum(as.numeric(unique(totUnitsPositive))), tot = max(totUnitsTested)),
    by = .(source, matrix, country)]
}
