% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE10.R
\name{tableE10}
\alias{tableE10}
\title{tableE10}
\usage{
tableE10(
  df_AMR = read_AMR(),
  year = 2024,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_AMR}{The isolate based data object}

\item{year}{the year to filter}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Produce Table 10 in Annex E: Occurrence of resistance (%) to selected
antimicrobials in methicillin-resistant Staphylococcus aureus from
food and animals, 2024.
}
