% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE12.R
\name{tableE12}
\alias{tableE12}
\title{tableE12}
\usage{
tableE12(
  df_AMR = read_AMR(),
  year = 2024,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_AMR}{The isolate based data object}

\item{year}{the year to filter}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Produce Annex E - Table 12: MDR patterns in methicillin-resistant
Staphylococcus aureus isolates from animals and food, 2024.
}
