% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE3_E4.R
\name{tableE3_E4_inner}
\alias{tableE3_E4_inner}
\title{tableE3_E4_inner}
\usage{
tableE3_E4_inner(
  df_prev = read_prev(),
  year = NULL,
  path_csv = tempfile(fileext = ".csv")
)
}
\value{
a path to a csv file
}
\description{
tableE3_E4_inner
}
