% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE5_E6.R
\name{tableE5_E6_inner}
\alias{tableE5_E6_inner}
\title{tableE5_E6}
\usage{
tableE5_E6_inner(
  df_prev = read_prev(),
  year = NULL,
  food = FALSE,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{food}{logical if you want food (TRUE) or nonfood (FALSE)}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce a table
}
