% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE5_E6_E7.R
\name{tableE7}
\alias{tableE7}
\title{tableE7}
\usage{
tableE7(
  df_prev = prev_by_samplingID(),
  year = 2023,
  food = FALSE,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{food}{logical if you want food (TRUE) or nonfood (FALSE)}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce Annex E table 7: Methicillin-resistant Staphylococcus
aureus in non-food-producing animals, clinical investigations,
2023.
}
