% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{figure2}
\usage{
figure2(df_prev = read_prev(), years = c(2023, 2024))
}
\arguments{
\item{df_prev}{The data object}

\item{years}{the years to filter}
}
\value{
A plot
}
\description{
Produce Figure 2 in the main chapter
}
