% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_2.R
\name{table1}
\alias{table1}
\title{table1}
\usage{
table1(df_prev = read_prev(), path_csv = tempfile(fileext = ".csv"))
}
\arguments{
\item{df_prev}{The object read_prev function}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Table 1 - CC398 only in main chapter
Produce Table 1 in the main chapter
}
