% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE2.R
\name{tableE2}
\alias{tableE2}
\title{tableE2}
\usage{
tableE2(
  df_prev = read_prev(),
  year = 2023,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce Annex E Table 2: Methicillin-resistant Staphylococcus
aureus in food, 2023.
}
