% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE3_E4.R
\name{tableE3_E4_inner}
\alias{tableE3_E4_inner}
\title{tableE3_E4_inner}
\usage{
tableE3_E4_inner(
  df_prev = prev_by_samplingID(),
  year = NULL,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to csv file}
}
\value{
a path to a csv file
}
\description{
Produce a table
}
